Complete list of book less than 0

+++ CodeFirstDB/ViewModle/Vote.cs
@@ -33,7 +33,7 @@ namespace CodeFirstDB.ViewModle
 
         [Column("Value")]
         [Required]
-        public Int64 Value { get; set; }
+        public int Value { get; set; }
     }
 
 }
